+++ src/deps/utf8_lite.h
@@ -25,6 +25,6 @@ bool ucs2_to_ucs4(uint16_t u2, uint32_t *u4);
 // 转换字符串 ucs4 -> ucs2 (末尾补\0)
 bool ucs4str_to_ucs2(uint32_t *ucs4str, int ucs4str_len, uint16_t *buf, bool ignore);
 // 转换字符串 ucs2 -> ucs4 (末尾补\0)
-bool ucs2str_to_ucs4(uint32_t *ucs2str, int ucs2str_len, uint16_t *buf);
+bool ucs2str_to_ucs4(uint16_t *ucs2str, int ucs2str_len, uint32_t *buf);
 
 #endif
